Drop new_data_module
