+++ flash/audio/speech_recognition/data.py
@@ -22,10 +22,10 @@ from flash.audio.speech_recognition.input import (
     SpeechRecognitionPathsInput,
 )
 from flash.audio.speech_recognition.output_transform import SpeechRecognitionOutputTransform
+from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
 from flash.core.data.input_transform import INPUT_TRANSFORM_TYPE, InputTransform
 from flash.core.data.io.input import Input
-from flash.core.data.new_data_module import DataModule
 from flash.core.registry import FlashRegistry
 from flash.core.utilities.stages import RunningStage
