+++ flash/image/classification/data.py
@@ -20,10 +20,10 @@ from torch.utils.data import Dataset
 
 from flash.core.data.base_viz import BaseVisualization
 from flash.core.data.callback import BaseDataFetcher
+from flash.core.data.data_module import DataModule, DatasetInput
 from flash.core.data.data_pipeline import DataPipelineState
 from flash.core.data.input_transform import INPUT_TRANSFORM_TYPE
 from flash.core.data.io.input import DataKeys, Input
-from flash.core.data.new_data_module import DataModule, DatasetInput
 from flash.core.data.utilities.paths import PATH_TYPE
 from flash.core.integrations.labelstudio.input import _parse_labelstudio_arguments, LabelStudioImageClassificationInput
 from flash.core.registry import FlashRegistry
