+++ flash/image/face_detection/data.py
@@ -15,9 +15,9 @@ from typing import Any, Dict, Optional, Sequence, Type
 
 from torch.utils.data import Dataset
 
+from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
 from flash.core.data.io.input import Input
-from flash.core.data.new_data_module import DataModule
 from flash.core.utilities.stages import RunningStage
 from flash.core.utilities.types import INPUT_TRANSFORM_TYPE
 from flash.image.classification.data import ImageClassificationFilesInput, ImageClassificationFolderInput
