+++ flash/image/style_transfer/data.py
@@ -16,9 +16,9 @@ from typing import Any, Collection, Dict, Optional, Sequence, Type
 import numpy as np
 import torch
 
+from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
 from flash.core.data.io.input import Input
-from flash.core.data.new_data_module import DataModule
 from flash.core.utilities.stages import RunningStage
 from flash.core.utilities.types import INPUT_TRANSFORM_TYPE
 from flash.image.classification.input import ImageClassificationFilesInput, ImageClassificationFolderInput
