+++ flash/text/seq2seq/core/data.py
@@ -13,9 +13,9 @@
 # limitations under the License.
 from typing import Any, Dict, List, Optional, Type, Union
 
+from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
 from flash.core.data.io.input import Input
-from flash.core.data.new_data_module import DataModule
 from flash.core.data.utilities.paths import PATH_TYPE
 from flash.core.integrations.transformers.input_transform import TransformersInputTransform
 from flash.core.utilities.imports import _TEXT_AVAILABLE
