+++ flash_examples/flash_components/custom_data_loading.py
@@ -23,10 +23,10 @@ from PIL import Image
 from pytorch_lightning import seed_everything
 
 from flash import _PACKAGE_ROOT, RunningStage
+from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
 from flash.core.data.input_transform import INPUT_TRANSFORM_TYPE, InputTransform
 from flash.core.data.io.input import DataKeys, Input
-from flash.core.data.new_data_module import DataModule
 from flash.core.data.utils import download_data
 
 seed_everything(42)
