+++ tests/core/data/test_callbacks.py
@@ -17,9 +17,9 @@ import torch
 from torch import tensor
 
 from flash.core.data.callback import BaseDataFetcher
+from flash.core.data.data_module import DataModule
 from flash.core.data.input_transform import InputTransform
 from flash.core.data.io.input import Input
-from flash.core.data.new_data_module import DataModule
 from flash.core.utilities.stages import RunningStage
